Added audio frame processing signal.
Audio frames won't process with the same frequency of process or physics_process. So it's important to know when a new frame is ready. Deferred function call is used to avoid multithreading problems. Optimization can be done if needed.

TODO :
The clarity values of the peaks are mostly useless. they get swapped with volume values with the spectrum analyzer. While the game currently uses both at the same time it's possible to join the both into a single audio effect. then the signal should give correct volumes and it should also give a total volume for the entire spectrum as a second argument in the emitted signal.